clutter: Change stage view scale to be float
To support fractional scaling, change the stage view scale to be a
float instead of an int. Also change the places where it is retrieved
and used when scaling things.

https://bugzilla.gnome.org/show_bug.cgi?id=765011

+++ clutter/clutter/clutter-stage-view.h
@@ -61,7 +61,7 @@ void clutter_stage_view_blit_offscreen (ClutterStageView            *view,
 					const cairo_rectangle_int_t *clip);
 
 CLUTTER_AVAILABLE_IN_MUTTER
-int clutter_stage_view_get_scale (ClutterStageView *view);
+float clutter_stage_view_get_scale (ClutterStageView *view);
 
 gboolean clutter_stage_view_is_dirty_viewport (ClutterStageView *view);
